requirements.txt
python-gnupg
clipboard
selenium
docopt
gitpython
